requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
django-admin-relation-links==0.2.5
django-mysql==4.5.0
drf-yasg==1.20.0
swagger-spec-validator==2.7.4